First version of the image support
Using DataContainer (buffered) to store the image data in the plate model (slows down parsing, but keeps the view slim).
The WellViewer got two constructors for the tooltip and the view. using TableView to display the image data.
Adding small information window to the HeatWell to informa about the launching of the WellViewer.
Changing the image data parsing in the HeatMapViewerNodeModel.