Converting python and matlab server from maven to eclipse projects
this makes it easier to compile from eclipes.
Also when importing the projects into the workspace, there were always
problems with the maven dependencies. For instance there is no stable
public repository for the cajo-library so far (I'm not sure if it is
still developped). So Here we resolve to just have a static library
without proper dependency mangement (but I guess this is barable since
there are only few dependencies anyway).

Also I added a TODO file.