cosmetics in Matlab Local
changed the name session to thread (so not to be confused with session, meaning an instance of a running matlab application) and it’s not a node either (since the node id is not accessible). It’s not elegant but counting threads is at least a way to distinguish different threads accessing  on the matlab application.
Furthermore I removed the line printing to the matlab console (this will be done from the node).