Adjust the “Open in Matlab” nodes
these nodes now use an instance of MatlabLocal to launch and communicate with matlab. Hence the platform dependent code (only for Mac OSx) to launch matlab was removed. The corresponding TODO was removed.
The option to launch a new Matlab instance each time the node is executed was removed from the node dialog.
The javadoc was completed and comments added where needed.
The Matlab script to load and convert the hashmap binary from within matlab was adjusted (essentially the message that was displayed in the command window).