Add MATLAB data type property in the KNIME preference dialog
The datatype was an open before, but only in the OpenInMatlab node. There it has been removed from the node dialog.
Instead there is a new combobox in the preference tab of the KNIME scripting integration.

There was a small bug in the plot code generation (MatlabCode); a new line was missing before the print statement to save the plot. this is important since if the snippet has as last line a % character, this statement would end up as a comment.

The hashmaputils.m script had the following bug; the option for the type are {map, dataset, struct} but the class-names are {containers.Map, dataset, struc}… careful!
Furthermore the additional arguments (varargin) were already made obsolete during the modifications of last commit.
Finally the docs were a bit modified to better reflect the current situation.