Add more stringent cell type handling for MATLAB scripting nodes input.
Add an exception type that then allows to uses the node logger to show a decent error message in the KNIME console.
The type checking in the tables conversion of MatlabTable was too relaxed. Astonishingly the ImgPlusCell from the KNIME Image Processing yielded true if its DataColumnSpec was interrogated for StringCell compatibility…