Prepare the Remote implementation of the MATLAB server
Change the AbstractMatlabScriptingNodeModel to handle all of the necessairy preference dialog properties. Propagate the parameters through the classes.

Add a second data transfer method:
The first method was file-based. This second one is an adaptation of the way it was working in the old server. In the re-implemenation I took care that all the Matlab code is generated in the MatlabCode class. Also we don’t load all the data in memory (as before) but we push and pull one table line after another (this is the way KNIME imposes the access to the BufferedDataTable).
This method seems to be a bit slower, but is more efficient on the memory side and does not weight on the disk space.

Change the hashmaputils.m variable names and update documentation

Define Interface for the server classes.