Restructure the MATLAB server and the client for the remote access. (NOT QUITE FINISHED MatlabCode needs some revision)
The way the client was interacting with the proxy on the remote server did not work at all, since it is not possible to get an object from the remote JVM and then manipulate it as if the client would “be” on the remote machine (see cajo docs)
So now the MatlabServer also implements the MatlabOperations interface from the MatlabControl library. This is to forward all the calls to MATLAB to the server and then get back static content.
The development of the task-methods in of the MatlabClient-Remote is complete but not yet tested.

The file manipulations in the MatlabCode class were removed. The method function signatures have become a bit more complex, but on the other hand what they do has become more transparent. Since this class has outgrown itself this re-organisation has become necessary. Some of the manipulations are done now in the MatlabFileTransfer class, which has added functionality to accommodate transfer of data for processing on the local machine as well as the remote host.