Add a second operation mode to the MatlabController
this is an attempt to resolve the jmi connection issues I experienced when the MatlabController was used by the MatlabServer. Somehow it is possible to start MATLAB during instantiation but not if the connect() method is called during an acquireMatlabProxy call.

The second operating mode opens the MATLAB application immediately when the server is created and tries to reconnect immediately in the callback of the disconnection callback.
The initial mode (lazy) is used for running MATLAB on the local machine.

Change the stdout to some more meaningful messages.