Do a better singleton implementation of the MatlabController
the controller is now a singleton that holds an ArrayBlockingQueue to
manage the access to the Matlab instances. The constructor was
abolished and replaced by a getInstance() method. The parameters of the
MatlabController are now set via a spearate method.
The deletion of the parser file (hashmaputils.m) was commented out,
since it caused problems (one process thinks that it created it and
before it can use it another node deleted it)