Make the column Names more persistent for Matlab snippets
For tables the same strategy is applied as for the datasets (putting the column header in the object properties).
For containers.Map and struct, the header stays the same if the it is untouched. Once the header is modified, variable names from Matlab have to be taken as headers (spaces removed and such) in oder to avoid confusion and ambiguities.